+++ CHANGELOG.rst
@@ -2,8 +2,30 @@ This document contains the ``fsleyes-widgets`` release history in reverse
 chronological order.
 
 
+0.11.0 (Thursday February 18th 2021)
+------------------------------------
 
-0.10.0 (Wednesday February 10th 2020)
+
+Added
+^^^^^
+
+
+* New functions for querying the environment at runtime, including,
+  :func:`.wxVersion` (not to be confused with the deprecated
+  :func:`.wxversion`), :func:`.wxPlatform`, :func:`.wxFlavour`, :func:`.frozen`,
+  :func:`.canHaveGui`, :func:`.haveGui`, :func:`.inSSHSession`,
+  :func:`.inVNCSession`, :func:`.glVersion`, :func:`.glRenderer`, and
+  :func:`.glIsSoftwareRenderer`.
+
+
+Deprecated
+^^^^^^^^^^
+
+
+* The :func:`.wxversion` function has been replaced by :func:`wxFlavour`.
+
+
+0.10.0 (Wednesday February 10th 2021)
 -------------------------------------